refactor: Extract RoleModal component for role assignment display
